Initial project prototype
This is the first, very rough-cut implementation.

I'm not happy with the interaction of robot => surface => location as it's a
bit unwieldy. I used a movement strategy for the robot, except robot has
no behaviour outside of that, rendering it somewhat pointless.

I'm considering having placed vs unplaced robot classes given the almost
complete absence of shared behaviour.